+++ src/DataConnector/DataConnector_oci.php
@@ -995,25 +995,29 @@ class DataConnector_oci extends DataConnector
      */
     public function loadPlatformNonce($nonce)
     {
+        if (parent::useMemcache()) {
+            $ok = parent::loadPlatformNonce($nonce);
+        } else {
 // Delete any expired nonce values
-        $now = date("{$this->dateFormat} {$this->timeFormat}", time());
-        $sql = "DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE expires <= :now';
-        $query = oci_parse($this->db, $sql);
-        oci_bind_by_name($query, 'now', $now);
-        $this->executeQuery($sql, $query);
+            $now = date("{$this->dateFormat} {$this->timeFormat}", time());
+            $sql = "DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE expires <= :now';
+            $query = oci_parse($this->db, $sql);
+            oci_bind_by_name($query, 'now', $now);
+            $this->executeQuery($sql, $query);
 
 // Load the nonce
-        $id = $nonce->getPlatform()->getRecordId();
-        $value = $nonce->getValue();
-        $sql = "SELECT value T FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = :id) AND (value = :value)';
-        $query = oci_parse($this->db, $sql);
-        oci_bind_by_name($query, 'id', $id);
-        oci_bind_by_name($query, 'value', $value);
-        $ok = $this->executeQuery($sql, $query, false);
-        if ($ok) {
-            $row = oci_fetch_assoc($query);
-            if ($row === false) {
-                $ok = false;
+            $id = $nonce->getPlatform()->getRecordId();
+            $value = $nonce->getValue();
+            $sql = "SELECT value T FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = :id) AND (value = :value)';
+            $query = oci_parse($this->db, $sql);
+            oci_bind_by_name($query, 'id', $id);
+            oci_bind_by_name($query, 'value', $value);
+            $ok = $this->executeQuery($sql, $query, false);
+            if ($ok) {
+                $row = oci_fetch_assoc($query);
+                if ($row === false) {
+                    $ok = false;
+                }
             }
         }
 
@@ -1029,15 +1033,19 @@ class DataConnector_oci extends DataConnector
      */
     public function savePlatformNonce($nonce)
     {
-        $id = $nonce->getPlatform()->getRecordId();
-        $value = $nonce->getValue();
-        $expires = date("{$this->dateFormat} {$this->timeFormat}", $nonce->expires);
-        $sql = "INSERT INTO {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' (consumer_pk, value, expires) VALUES (:id, :value, :expires)';
-        $query = oci_parse($this->db, $sql);
-        oci_bind_by_name($query, 'id', $id);
-        oci_bind_by_name($query, 'value', $value);
-        oci_bind_by_name($query, 'expires', $expires);
-        $ok = $this->executeQuery($sql, $query);
+        if (parent::useMemcache()) {
+            $ok = parent::savePlatformNonce($nonce);
+        } else {
+            $id = $nonce->getPlatform()->getRecordId();
+            $value = $nonce->getValue();
+            $expires = date("{$this->dateFormat} {$this->timeFormat}", $nonce->expires);
+            $sql = "INSERT INTO {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' (consumer_pk, value, expires) VALUES (:id, :value, :expires)';
+            $query = oci_parse($this->db, $sql);
+            oci_bind_by_name($query, 'id', $id);
+            oci_bind_by_name($query, 'value', $value);
+            oci_bind_by_name($query, 'expires', $expires);
+            $ok = $this->executeQuery($sql, $query);
+        }
 
         return $ok;
     }
@@ -1051,13 +1059,17 @@ class DataConnector_oci extends DataConnector
      */
     public function deletePlatformNonce($nonce)
     {
-        $sql = "DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = :id) AND (value = :value)';
-        $query = oci_parse($this->db, $sql);
-        $id = $nonce->getPlatform()->getRecordId();
-        oci_bind_by_name($query, 'id', $id);
-        $value = $nonce->getValue();
-        oci_bind_by_name($query, 'value', $value);
-        $ok = $this->executeQuery($sql, $query);
+        if (parent::useMemcache()) {
+            $ok = parent::deletePlatformNonce($nonce);
+        } else {
+            $sql = "DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = :id) AND (value = :value)';
+            $query = oci_parse($this->db, $sql);
+            $id = $nonce->getPlatform()->getRecordId();
+            oci_bind_by_name($query, 'id', $id);
+            $value = $nonce->getValue();
+            oci_bind_by_name($query, 'value', $value);
+            $ok = $this->executeQuery($sql, $query);
+        }
 
         return $ok;
     }
