+++ src/DataConnector/DataConnector_pg.php
@@ -787,20 +787,24 @@ class DataConnector_pg extends DataConnector
      */
     public function loadPlatformNonce($nonce)
     {
-        $ok = false;
+        if (parent::useMemcache()) {
+            $ok = parent::loadPlatformNonce($nonce);
+        } else {
+            $ok = false;
 
 // Delete any expired nonce values
-        $now = date("{$this->dateFormat} {$this->timeFormat}", time());
-        $sql = "DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . " WHERE expires <= '{$now}'";
-        $this->executeQuery($sql);
+            $now = date("{$this->dateFormat} {$this->timeFormat}", time());
+            $sql = "DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . " WHERE expires <= '{$now}'";
+            $this->executeQuery($sql);
 
 // Load the nonce
-        $sql = sprintf("SELECT value AS T FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = %d) AND (value = %s)',
-            $nonce->getPlatform()->getRecordId(), $this->escape($nonce->getValue()));
-        $rsNonce = $this->executeQuery($sql, false);
-        if ($rsNonce) {
-            if (pg_fetch_object($rsNonce)) {
-                $ok = true;
+            $sql = sprintf("SELECT value AS T FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = %d) AND (value = %s)',
+                $nonce->getPlatform()->getRecordId(), $this->escape($nonce->getValue()));
+            $rsNonce = $this->executeQuery($sql, false);
+            if ($rsNonce) {
+                if (pg_fetch_object($rsNonce)) {
+                    $ok = true;
+                }
             }
         }
 
@@ -816,10 +820,14 @@ class DataConnector_pg extends DataConnector
      */
     public function savePlatformNonce($nonce)
     {
-        $expires = date("{$this->dateFormat} {$this->timeFormat}", $nonce->expires);
-        $sql = sprintf("INSERT INTO {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . " (consumer_pk, value, expires) VALUES (%d, %s, %s)",
-            $nonce->getPlatform()->getRecordId(), $this->escape($nonce->getValue()), $this->escape($expires));
-        $ok = $this->executeQuery($sql);
+        if (parent::useMemcache()) {
+            $ok = parent::savePlatformNonce($nonce);
+        } else {
+            $expires = date("{$this->dateFormat} {$this->timeFormat}", $nonce->expires);
+            $sql = sprintf("INSERT INTO {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . " (consumer_pk, value, expires) VALUES (%d, %s, %s)",
+                $nonce->getPlatform()->getRecordId(), $this->escape($nonce->getValue()), $this->escape($expires));
+            $ok = $this->executeQuery($sql);
+        }
 
         return $ok;
     }
@@ -833,9 +841,13 @@ class DataConnector_pg extends DataConnector
      */
     public function deletePlatformNonce($nonce)
     {
-        $sql = sprintf("DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = %d) AND (value = %s)',
-            $nonce->getPlatform()->getRecordId(), $this->escape($nonce->getValue()));
-        $ok = $this->executeQuery($sql);
+        if (parent::useMemcache()) {
+            $ok = parent::deletePlatformNonce($nonce);
+        } else {
+            $sql = sprintf("DELETE FROM {$this->dbTableNamePrefix}" . static::NONCE_TABLE_NAME . ' WHERE (consumer_pk = %d) AND (value = %s)',
+                $nonce->getPlatform()->getRecordId(), $this->escape($nonce->getValue()));
+            $ok = $this->executeQuery($sql);
+        }
 
         return $ok;
     }
